refactor(transport): moved APIError to separate package to avoid import cycles
