+++ http/transport.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/weaveworks/flux"
 	"github.com/weaveworks/flux/api"
+	"github.com/weaveworks/flux/http/error"
 	"github.com/weaveworks/flux/http/websocket"
 	"github.com/weaveworks/flux/jobs"
 	"github.com/weaveworks/flux/platform"
@@ -76,34 +77,6 @@ func NewHandler(s api.FluxService, r *mux.Router, logger log.Logger, h *stdprome
 		RouteMatcher: r,
 		Duration:     h,
 	}.Wrap(r)
-}
-
-// When an API call fails, we may want to distinguish among the causes
-// by status code. This type can be used as the base error when we get
-// a non-"HTTP 20x" response, retrievable with errors.Cause(err).
-type APIError struct {
-	StatusCode int
-	Status     string
-	Body       string
-}
-
-func (err *APIError) Error() string {
-	return fmt.Sprintf("%s (%s)", err.Status, err.Body)
-}
-
-// Does this error mean the API service is unavailable?
-func (err *APIError) IsUnavailable() bool {
-	switch err.StatusCode {
-	case 502, 503, 504:
-		return true
-	}
-	return false
-}
-
-// Is this API call missing? This usually indicates that there is a
-// version mismatch between the client and the service.
-func (err *APIError) IsMissing() bool {
-	return err.StatusCode == http.StatusNotFound
 }
 
 // The idea here is to place the handleFoo and invokeFoo functions next to each
@@ -845,7 +818,7 @@ func executeRequest(client *http.Client, req *http.Request) (*http.Response, err
 	default:
 		buf, _ := ioutil.ReadAll(resp.Body)
 		body := strings.TrimSpace(string(buf))
-		return nil, &APIError{
+		return nil, &httperror.APIError{
 			StatusCode: resp.StatusCode,
 			Status:     resp.Status,
 			Body:       body,
